Conver to base focus for now
