+++ remake/src/app/components/FormAddressHelpers/SelectBlockchain.tsx
@@ -17,7 +17,8 @@ const SelectBlockchain = () => {
     setBlockchain(value)
   }
 
-  const blockchains = ["Solana", "Ethereum", "Base"]
+  // const blockchains = ["Solana", "Ethereum", "Base"]
+  const blockchains = ["Base"]
     
   return (
     <>
