+++ remake/src/app/layout.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
 import NavBar from "./components/NavBar";
 import { WalletProvider } from "./context"
 import Socials from "./components/Socials"
-
-const inter = Inter({ subsets: ["latin"] });
+// import Head from "next/head"
+// import Script from "next/script"
 
 export const metadata: Metadata = {
   title: "Wallet Finance",
@@ -20,8 +19,20 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
+      {/* <Head> */}
+      {/*  Google Tag Manager */}
+        {/* <Script async src='https://www.googletagmanager.com/gtm.js?id='/> */}
+          
+        {/* <Script dangerouslySetInnerHTML={
+          function(w: any,d: any,s: any,l: any,i: any){w[l]=w[l]||[];w[l].push({'gtm.start':
+            new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+            j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.af+i+dl;f.parentNode.insertBefore(j,f);
+            }(window,document,'script','dataLayer','GTM-WF7HQMKT');
+        }/> */}
+        {/* End Google Tag Manager */}
+      {/* </Head> */}
       <WalletProvider>
-      <body className={inter.className}>
+      <body style={{height: "95vh"}}>
           <header>
             <NavBar /> 
           </header>
